Delete useless interfaces, extract ManifestEntry to it's own class (and rename to ModEntry) instead of being a sub-class
